+++ apps/backend/src/modules/chat-messages/libs/types/types.ts
@@ -4,4 +4,5 @@ export {
 	type ChatMessageItemWithReceiverIdResponseDto,
 	type ChatMessageUpdateRequestDto,
 	type ReadChatMessagesRequestDto,
+	type ReadChatMessagesResponseDto,
 } from "@trackmates/shared";