+++ apps/backend/src/modules/permissions/permissions.ts
@@ -14,8 +14,8 @@ const permissionController = new PermissionController(
 
 export { permissionController, permissionService };
 export {
-	type AllPermissionsResponseDto,
 	type PermissionResponseDto,
+	type PermissionsGetAllResponseDto,
 } from "./libs/types/types.js";
 export { PermissionEntity } from "./permission.entity.js";
 export { PermissionModel } from "./permission.model.js";