+++ apps/frontend/src/modules/user-notifications/slices/user-notifications.slice.ts
@@ -1,9 +1,18 @@
-import { createSlice } from "@reduxjs/toolkit";
+import { type PayloadAction, createSlice } from "@reduxjs/toolkit";
 
-import { DataStatus } from "~/libs/enums/enums.js";
+import {
+	DataStatus,
+	NotificationFilter,
+	NotificationStatus,
+} from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
-import { type NotificationResponseDto } from "../libs/types/types.js";
+import { NEW_NOTIFICATION_COUNT } from "../libs/constants/constants.js";
+import { notificationFilterToType } from "../libs/maps/maps.js";
+import {
+	type NotificationResponseDto,
+	type ReadNotificationsResponseDto,
+} from "../libs/types/types.js";
 import {
 	getUnreadNotificationsCount,
 	getUserNotifications,
@@ -12,12 +21,14 @@ import {
 
 type State = {
 	dataStatus: ValueOf<typeof DataStatus>;
+	notificationType: ValueOf<typeof NotificationFilter>;
 	notifications: NotificationResponseDto[];
 	unreadNotificationsCount: number;
 };
 
 const initialState: State = {
 	dataStatus: DataStatus.IDLE,
+	notificationType: NotificationFilter.ALL,
 	notifications: [],
 	unreadNotificationsCount: 0,
 };
@@ -44,6 +55,9 @@ const { actions, name, reducer } = createSlice({
 
 				return updatedNotification ?? value;
 			});
+			state.unreadNotificationsCount = Number(
+				action.payload.unreadNotificationsCount,
+			);
 			state.dataStatus = DataStatus.FULFILLED;
 		});
 		builder.addCase(setReadNotifications.rejected, (state) => {
@@ -59,7 +73,55 @@ const { actions, name, reducer } = createSlice({
 	},
 	initialState,
 	name: "user-notifications",
-	reducers: {},
+	reducers: {
+		deleteNotification(state, action: PayloadAction<NotificationResponseDto>) {
+			state.notifications = state.notifications.filter((notification) => {
+				return notification.id != action.payload.id;
+			});
+
+			if (action.payload.status === NotificationStatus.UNREAD) {
+				state.unreadNotificationsCount =
+					state.unreadNotificationsCount - NEW_NOTIFICATION_COUNT;
+			}
+		},
+		newNotification(state, action: PayloadAction<NotificationResponseDto>) {
+			const needUpdateNotifications =
+				state.notificationType === NotificationFilter.ALL ||
+				notificationFilterToType[state.notificationType] ===
+					action.payload.type;
+
+			if (needUpdateNotifications) {
+				state.notifications = [action.payload, ...state.notifications];
+			}
+
+			state.unreadNotificationsCount =
+				state.unreadNotificationsCount + NEW_NOTIFICATION_COUNT;
+		},
+		setNotificationType(
+			state,
+			action: PayloadAction<ValueOf<typeof NotificationFilter>>,
+		) {
+			state.notificationType = action.payload;
+			state.dataStatus = DataStatus.FULFILLED;
+		},
+		updateReadNotifications(
+			state,
+			action: PayloadAction<ReadNotificationsResponseDto>,
+		) {
+			state.notifications = state.notifications.map((value) => {
+				const updatedNotification = action.payload.items.find(
+					(notification) => {
+						return notification.id === value.id;
+					},
+				);
+
+				return updatedNotification ?? value;
+			});
+			state.unreadNotificationsCount = Number(
+				action.payload.unreadNotificationsCount,
+			);
+		},
+	},
 });
 
 export { actions, name, reducer };