+++ apps/frontend/src/pages/management/libs/components/table/libs/components/table-cell/table-cell.tsx
@@ -3,14 +3,14 @@ import { getValidClassNames } from "~/libs/helpers/helpers.js";
 import styles from "./styles.module.css";
 
 type Properties = {
-	centered?: boolean;
 	children: React.ReactNode;
+	isCentered?: boolean;
 	width?: "medium" | "narrow" | "wide";
 };
 
 const TableCell: React.FC<Properties> = ({
-	centered = false,
 	children,
+	isCentered = false,
 	width = "wide",
 }: Properties) => {
 	return (
@@ -20,7 +20,7 @@ const TableCell: React.FC<Properties> = ({
 			<div
 				className={getValidClassNames(
 					styles["content"],
-					centered && styles["centered"],
+					isCentered && styles["centered"],
 				)}
 			>
 				{children}