+++ apps/frontend/src/pages/notifications/libs/enums/enums.ts
@@ -1,1 +1,0 @@
-export { QueryParameterName } from "./query-parameter-name.enum.js";