+++ packages/shared/src/index.ts
@@ -78,6 +78,7 @@ export {
 	ChatMessagesApiPath,
 	MessageStatus,
 	type ReadChatMessagesRequestDto,
+	type ReadChatMessagesResponseDto,
 	chatMessageCreateValidationSchema,
 	chatMessageIdParameterValidationSchema,
 	chatMessageUpdateValidationSchema,
@@ -149,12 +150,12 @@ export {
 	friendIdParameterValidationSchema,
 } from "./modules/friends/friends.js";
 export {
-	type AllGroupsResponseDto,
 	GroupError,
 	GroupErrorMessage,
 	type GroupRequestDto,
 	type GroupResponseDto,
 	GroupsApiPath,
+	type GroupsGetAllResponseDto,
 	groupIdAndPermissionIdParametersValidationSchema,
 	groupIdAndUserIdParametersValidationSchema,
 	groupIdParameterValidationSchema,
@@ -168,10 +169,11 @@ export {
 	type NotificationResponseDto,
 	NotificationStatus,
 	NotificationType,
+	type ReadNotificationsResponseDto,
 	type UpdateNotificationRequestDto,
+	notificationFilterToType,
 } from "./modules/notifications/notifications.js";
 export {
-	type AllPermissionsResponseDto,
 	PermissionError,
 	PermissionErrorMessage,
 	PermissionKey,
@@ -179,6 +181,7 @@ export {
 	type PermissionRequestDto,
 	type PermissionResponseDto,
 	PermissionsApiPath,
+	type PermissionsGetAllResponseDto,
 	permissionIdParameter,
 } from "./modules/permissions/permissions.js";
 export {