+++ packages/shared/src/modules/groups/libs/enums/group-error-message.enum.ts
@@ -4,7 +4,10 @@ const GroupErrorMessage = {
 	DELETE_GROUP_FORBIDDEN: "Delete group forbidden.",
 	GROUP_ALREADY_EXISTS: "Group already exists.",
 	GROUP_ALREADY_HAS_THIS_PERMISSION: "Group already has this permission.",
+	GROUP_CHANGE_FORBIDDEN: "Group change forbidden.",
+	GROUP_DELETION_FORBIDDEN: "Group deletion forbidden.",
 	GROUP_NOT_FOUND: "Group not found.",
+	PERMISSION_CHANGE_FORBIDDEN: "Permission change forbidden.",
 	USER_ALREADY_IN_GROUP: "User already in group.",
 } as const;