+++ packages/shared/src/modules/groups/libs/types/types.ts
@@ -1,3 +1,3 @@
-export { type AllGroupsResponseDto } from "./all-groups-response-dto.type.js";
 export { type GroupRequestDto } from "./group-request-dto.type.js";
 export { type GroupResponseDto } from "./group-response-dto.type.js";
+export { type GroupsGetAllResponseDto } from "./groups-get-all-response-dto.type.js";